+++ src/main/java/br/gov/model/cadastro/ImovelEconomia.java
@@ -18,17 +18,13 @@ import javax.persistence.Table;
 @Entity
 @Table(name="imovel_economia", schema="cadastro")
 public class ImovelEconomia implements Serializable{
-
-	/**
-	 * 
-	 */
 	private static final long serialVersionUID = -5595073207698176858L;
 
 	@Id
 	@GeneratedValue(strategy=GenerationType.SEQUENCE, generator="SEQ_IMOVEL_ECONOMIA")
 	@SequenceGenerator(name="SEQ_IMOVEL_ECONOMIA", schema="cadastro", sequenceName="seq_imovel_economia", allocationSize=1)
 	@Column(name="imec_id")
-	private Long id;
+	private Integer id;
 
 	@Column(name="imec_dscomplementoendereco")
 	private String complementoEndereco;
@@ -43,7 +39,7 @@ public class ImovelEconomia implements Serializable{
 	private String numeroIptu;
 	
 	@Column(name="imec_nncontratoenergia")
-	private Long numeroCelpa;
+	private Integer numeroCelpa;
 	
 	@Column(name="imec_nnareaconstruida")
 	private BigDecimal areaConstruida;
@@ -60,11 +56,11 @@ public class ImovelEconomia implements Serializable{
 	
 	public ImovelEconomia(){}
 
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
 
@@ -100,11 +96,11 @@ public class ImovelEconomia implements Serializable{
 		this.numeroIptu = numeroIptu;
 	}
 
-	public Long getNumeroCelpa() {
+	public Integer getNumeroCelpa() {
 		return numeroCelpa;
 	}
 
-	public void setNumeroCelpa(Long numeroCelpa) {
+	public void setNumeroCelpa(Integer numeroCelpa) {
 		this.numeroCelpa = numeroCelpa;
 	}
