+++ src/test/java/br/gov/servicos/cobranca/CobrancaDocumentoItemRepositorioTest.java
@@ -34,9 +34,9 @@ public class CobrancaDocumentoItemRepositorioTest {
 	@ShouldMatchDataSet("cobranca_documento_item_expected.yml")
 	@Cleanup(phase = TestExecutionPhase.AFTER, strategy = CleanupStrategy.USED_ROWS_ONLY)
 	public void apagarCobrancasDasContas(){
-		List<Long> ids = new ArrayList<Long>();
-		ids.add(1L);
-		ids.add(4L);
+		List<Integer> ids = new ArrayList<Integer>();
+		ids.add(1);
+		ids.add(4);
 		repositorio.apagarItensCobrancaDasContas(ids);
 	}
 }
